refactor: extract color filters and tab rendering from tab-manager
